[Trusted Types] Disable sample clipping in TT CSP reports.
CSP reports clip the sample to 40 characters. TT reports prepend the sink name to the 40-char clipped sample. That part should not be counted towards the limit. Solve this by having TT handle the clipping, and CSP pass it through.

The sink name is a JS object name followed by a property name (Element.innerHTML), which increases the entropy of the collected sample only marginally.

Bug: 991195
Change-Id: Id7a44d815d27a6bd341187c1318ccae1fd234ef7