[css-grid] Fix line name positions after implicit tracks
When serializing grid-template-rows/columns of a grid container, we get
the list of used track sizes, and then insert line names at the
corresponding indices.

Before this patch, these indices were simply taken from the computed
value. However, leading implicit tracks appear in the used value but not
in the computed one. Therefore, we have to offset the indices so that
the line names appear at the proper positions.

BUG=790988

TEST=external/wpt/css/css-grid/parsing/grid-template-columns-computed-implicit-track.html
TEST=external/wpt/css/css-grid/parsing/grid-template-rows-computed-implicit-track.html

Change-Id: I778c73034c7ff9fc11a7d8a46a4042cfd1bcd1ae