[css-layout-api] Custom layouts with no children nodes
It was discovered that creating a custom layout for an element with
no children nodes caused a crash. This reason being that if the node
being laid out is inline and has no child nodes,
NGBlockNode::FirstChild() will return the inline node being laid out
as its first child.
https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/layout/ng/ng_block_node.cc?sq=package:chromium&g=0&l=744&ssfr=1

Calling child.GetCustomLayoutChild() on this element failed because
GetCustomLayoutChild() is expecting block level nodes rather than
inline nodes.

Thus, in the case of custom layouts, we don't want to treat the
element being laid out as a child. To fix this, we must check that
the child returned from NGBlockNode::FirstChild() is not the same as
the node being laid out when collecting children for custom layouts.

Bug: 1038501
Change-Id: I0816dab025f70613634a7c968f3d8abfe46ae01c