Change EncryptionScheme to match latest spec
The current version of the EME spec
(https://rawgit.com/w3c/encrypted-media/master/index.html) has
encryptionScheme as a DOMString with an additional value allowed
("cbcs-1-9"). Updating the code to match this.

This also moves the feature to experimental so that "Enable experimental
web platform features" can be used to enable it.

With this change the W3C test
(https://wpt.live/encrypted-media/clearkey-check-encryption-scheme.https.html)
passes (when experimental features are enabled).

Bug: 1035153
Test: Updated media_blink_unittests, browser_tests, and WPT test pass
Change-Id: I00cb2ce8a4c4f52f95817471226b5e8129660247