Reduce report timeout in CSP reporting API WPT tests
In order to test that no report was sent, the WPT tests set a timeout
and check that no report was received once it expires. The current value
of the timeout is a bit too long, causing the test to sometimes timeout
in a flaky manner. This CL reduces the timeout and re-enables the test.

Bug: 850170
Change-Id: Ibc39291fa0bcfafa94cd84233b78973d74afff1c